quran-shorts-bot/requirements.txt
moviepy==1.0.3
Pillow==10.4.0
requests==2.31.0
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
numpy==1.26.4
